requirements.txt
# Core dependencies
streamlit==1.31.0
pdfplumber==0.10.3
pymupdf==1.23.8
sentence-transformers==2.3.1
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
pillow==10.2.0
typing-extensions==4.9.0

# OpenAI Agents SDK (NEW - Released Nov 20, 2025)
openai-agents==0.6.1
openai>=1.12.0